+++ share/cos/signal.go
@@ -1,3 +1,5 @@
+//+build !windows
+
 package cos
 
 import (
@@ -26,4 +28,21 @@ func GoStats() {
 			runtime.NumGoroutine(),
 			sizestr.ToString(int64(memStats.Alloc)))
 	}
+}
+
+//AfterSignal returns a channel which will be closed
+//after the given duration or until a SIGHUP is received
+func AfterSignal(d time.Duration) <-chan struct{} {
+	ch := make(chan struct{})
+	go func() {
+		sig := make(chan os.Signal, 1)
+		signal.Notify(sig, syscall.SIGHUP)
+		select {
+		case <-time.After(d):
+		case <-sig:
+		}
+		signal.Stop(sig)
+		close(ch)
+	}()
+	return ch
 }
