+++ share/cos/signal_windows.go
@@ -1,0 +1,20 @@
+//+build windows
+
+package cos
+
+import (
+	"time"
+)
+
+func GoStats() {
+	//noop
+}
+
+func AfterSignal(d time.Duration) <-chan struct{} {
+	ch := make(chan struct{})
+	go func() {
+		<-time.After(d)
+		close(ch)
+	}()
+	return ch
+}
