--- share/cos/sleep_windows.go
@@ -1,10 +1,0 @@
-//+build windows
-
-package cos
-
-import "time"
-
-//Sleep unless Signal
-func SleepSignal(d time.Duration) {
-	time.Sleep(d) //not supported
-}
